+++ CHANGELOG.md
@@ -11,6 +11,7 @@ LynxKite:
  - More compact UI for scalars and attributes.
  - Fill full width if only one project is open.
  - New _"Count most common"_ local aggregator added.
+ - Scalable _"Count distinct"_ and _"Most common"_ aggregators.
 
 ### 1.9.6